# specs/openapi.app.json
{
  "components": {
    "schemas": {
...
      "ActionResult": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "removed": {
            "title": "Removed",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "removed"
        ],
        "title": "ActionResult",
        "type": "object"
      },
...
      "BulkAction": {
        "enum": [
          "DELETE",
          "OPT_OUT"
        ],
        "title": "BulkAction",
        "type": "string"
      },
      "BulkRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/BulkAction"
          },
          "contactIds": {
            "items": {
              "type": "string"
            },
            "title": "Contactids",
            "type": "array"
          }
        },
        "required": [
          "action",
          "contactIds"
        ],
        "title": "BulkRequest",
        "type": "object"
      },
...
      "CleanUpAction": {
        "enum": [
          "INVALID",
          "OPTED_OUT"
        ],
        "title": "CleanUpAction",
        "type": "string"
      },
      "CleanUpRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/CleanUpAction"
          }
        },
        "required": [
          "action"
        ],
        "title": "CleanUpRequest",
        "type": "object"
      },
      "Contact": {
        "properties": {
          "accountId": {
            "title": "Accountid",
            "type": "string"
          },
          "cf1": {
            "default": "",
            "title": "Cf1",
            "type": "string"
          },
          "cf2": {
            "default": "",
            "title": "Cf2",
            "type": "string"
          },
          "cf3": {
            "default": "",
            "title": "Cf3",
            "type": "string"
          },
          "cf4": {
            "default": "",
            "title": "Cf4",
            "type": "string"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "firstName": {
            "default": "",
            "title": "Firstname",
            "type": "string"
          },
          "lastName": {
            "default": "",
            "title": "Lastname",
            "type": "string"
          },
          "listId": {
            "title": "Listid",
            "type": "string"
          },
          "mobile": {
            "title": "Mobile",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "listId",
          "mobile",
          "updatedAt"
        ],
        "title": "Contact",
        "type": "object"
      },
      "ContactHit": {
        "properties": {
          "firstName": {
            "title": "Firstname",
            "type": "string"
          },
          "lastName": {
            "title": "Lastname",
            "type": "string"
          },
          "listId": {
            "title": "Listid",
            "type": "string"
          },
          "mobile": {
            "title": "Mobile",
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "listId",
          "mobile"
        ],
        "title": "ContactHit",
        "type": "object"
      },
      "ContactInput": {
        "properties": {
          "cf1": {
            "default": "",
            "title": "Cf1",
            "type": "string"
          },
          "cf2": {
            "default": "",
            "title": "Cf2",
            "type": "string"
          },
          "cf3": {
            "default": "",
            "title": "Cf3",
            "type": "string"
          },
          "cf4": {
            "default": "",
            "title": "Cf4",
            "type": "string"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "firstName": {
            "default": "",
            "title": "Firstname",
            "type": "string"
          },
          "lastName": {
            "default": "",
            "title": "Lastname",
            "type": "string"
          },
          "mobile": {
            "title": "Mobile",
            "type": "string"
          }
        },
        "required": [
          "mobile"
        ],
        "title": "ContactInput",
        "type": "object"
      },
      "ContactList": {
        "properties": {
          "contactCount": {
            "default": 0,
            "title": "Contactcount",
            "type": "integer"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ListKind",
            "default": "STANDARD"
          },
          "listId": {
            "title": "Listid",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "listId",
          "name"
        ],
        "title": "ContactList",
        "type": "object"
      },
      "ContactPage": {
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Contact"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ContactPage",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ImportReport": {
        "properties": {
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "imported",
          "message",
          "skipped",
          "updated"
        ],
        "title": "ImportReport",
        "type": "object"
      },
      "ImportRequest": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ContactInput"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "ImportRequest",
        "type": "object"
      },
...
      "ListKind": {
        "enum": [
          "OPT_OUT",
          "STANDARD"
        ],
        "title": "ListKind",
        "type": "string"
      },
      "ListRequest": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ListRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "txtlocal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/app/contacts/search": {
      "get": {
        "operationId": "search_api_app_contacts_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContactHit"
                  },
                  "title": "Response Search Api App Contacts Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Search"
      }
    },
...
    "/api/app/lists": {
      "get": {
        "operationId": "lists_api_app_lists_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContactList"
                  },
                  "title": "Response Lists Api App Lists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Lists"
      },
      "post": {
        "operationId": "create_list_api_app_lists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Create List"
      }
    },
    "/api/app/lists/{listId}": {
      "delete": {
        "operationId": "delete_list_api_app_lists__listId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Delete List"
      },
      "patch": {
        "operationId": "rename_list_api_app_lists__listId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Rename List"
      }
    },
    "/api/app/lists/{listId}/clean-up": {
      "post": {
        "operationId": "clean_up_api_app_lists__listId__clean_up_post",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Clean Up"
      }
    },
    "/api/app/lists/{listId}/contacts": {
      "get": {
        "operationId": "contacts_api_app_lists__listId__contacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Contacts"
      },
      "post": {
        "operationId": "add_contact_api_app_lists__listId__contacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Add Contact"
      }
    },
    "/api/app/lists/{listId}/contacts/bulk": {
      "post": {
        "operationId": "bulk_api_app_lists__listId__contacts_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Bulk"
      }
    },
    "/api/app/lists/{listId}/contacts/import": {
      "post": {
        "operationId": "import_contacts_api_app_lists__listId__contacts_import_post",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Import Contacts"
      }
    },
    "/api/app/lists/{listId}/contacts/{contactId}": {
      "delete": {
        "operationId": "remove_contact_api_app_lists__listId__contacts__contactId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "title": "Contactid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Remove Contact"
      },
      "patch": {
        "operationId": "update_contact_api_app_lists__listId__contacts__contactId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "title": "Contactid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Update Contact"
      }
    },
    "/api/app/lists/{listId}/export": {
      "get": {
        "operationId": "export_api_app_lists__listId__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "title": "Listid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Export"
      }
    },
...
  }
}